refactor: update ServiceProviderUserAuthToken to use ServiceProviderType enum
